ANSI'fied arch/i86/kernel/ directory, and verified that result compiles and boots.

+++ elks/arch/i86/kernel/irq.c
@@ -242,7 +242,7 @@ int request_irq(int irq, void (*handler)(), void *dev_id)
     action->handler = handler;
     action->dev_id = dev_id;
 
-    enable_irq(irq);
+    enable_irq((unsigned int) irq);
 
     restore_flags(flags);
